Convert chart responsiveness code intocustom hook for usability
