+++ components/charts/areabar.js
@@ -1,8 +1,5 @@
 "use client"
-import { useContext, useEffect, useRef } from 'react';
-import { navMountContext } from '@/app/dashboard/provider';
-import { navCollapseContext } from '@/app/dashboard/provider';
-
+import { useChartResponsive } from '@/hooks/responsive';
 
 import {
  Chart as ChartJS,
@@ -76,15 +73,8 @@ export const data = {
 
 
  const AreaChart = ( ) => {
-
-  const {navMount} = useContext(navMountContext)
-  const {navCollapse} = useContext(navCollapseContext)
-  const resizeRef = useRef(null)
-  
-  useEffect(()=>{ 
-   resizeRef.current.resize()
-  }, 
-   [navMount, navCollapse])
+  const [resizeRef ] = useChartResponsive()
+ 
 
   return (
    <div className='bg-white w-full' >
